+++ src/enums.js
@@ -635,7 +635,7 @@ export const NavigationMode = {
  * So small nodes make diagram fit into the screen space, but they have no details. Our solution is to show cursor and selected items
  * of the diagram in full size and draw all other as markers.
  *
- * This enumeration controls visibility of neighbours of the cursor node in the auto fit mode. It allows to draw 
+ * This enumeration controls visibility of neighbors of the cursor node in the auto fit mode. It allows to draw 
  * them in full size regardless of available space.
  *
  * @enum {NavigationMode}
@@ -646,7 +646,7 @@ export const NeighboursSelectionMode = {
      */
     ParentsAndChildren: 0,
     /**
-     * Selects parents, children, spouses and siblings of the cursor item.
+     * Selects parents, children, and siblings of the cursor item.
      */
     ParentsChildrenSiblingsAndSpouses: 1
 };